[+] Added all interpreters

+++ src/eu/securebit/gungame/game/GunGame.java
@@ -9,31 +9,86 @@ import org.bukkit.potion.PotionEffect;
 import org.bukkit.util.Vector;
 
 import eu.securebit.gungame.Main;
-import eu.securebit.gungame.framework.ActionInterface;
-import eu.securebit.gungame.framework.Settings;
+import eu.securebit.gungame.interpreter.GameOptions;
+import eu.securebit.gungame.interpreter.GunGameScoreboard;
+import eu.securebit.gungame.interpreter.LevelManager;
+import eu.securebit.gungame.interpreter.LocationManager;
+import eu.securebit.gungame.interpreter.Messanger;
+import eu.securebit.gungame.interpreter.impl.CraftGameOptions;
+import eu.securebit.gungame.interpreter.impl.CraftGunGameScoreboard;
+import eu.securebit.gungame.interpreter.impl.CraftLevelManager;
+import eu.securebit.gungame.interpreter.impl.CraftMessanger;
+import eu.securebit.gungame.io.configs.FileGameConfig;
+import eu.securebit.gungame.io.configs.FileLevels;
+import eu.securebit.gungame.io.configs.FileMessages;
+import eu.securebit.gungame.io.configs.FileOptions;
+import eu.securebit.gungame.io.configs.FileScoreboard;
+import eu.securebit.gungame.io.directories.RootDirectory;
 
-public abstract class GunGame extends CraftGame<GunGamePlayer> {
+public class GunGame extends CraftGame<GunGamePlayer> {
 
 	private GunGamePlayer winner;
+	
 	private GunGameScoreboard board;
+	private LevelManager levelManager;
+	private Messanger messanger;
+	private GameOptions options;
+	private LocationManager locationManager;
 	
-	private ActionInterface actionInterface;
-	private Settings settings;
+	private String name;
 	
-	public GunGame(Settings settings, ActionInterface actionInterface) {
+	private FileGameConfig config;
+	
+	private GameInterface gameInterface;
+	
+	public GunGame(FileGameConfig config, String name, GameInterface gameInterface) {
 		super(Main.instance());
 		
-		this.settings = settings;
-		this.actionInterface = actionInterface;
-		this.board = new GunGameScoreboard(this, this.getSettings().getUUID());
+		this.name = name;
+		this.gameInterface = gameInterface;
 		
-		if (this.board.isEnabled()) {
-			if (this.board.exists()) {
-				this.board.delete();
-			}
-			
-			this.board.create();
+		RootDirectory root = Main.instance().getRootDirectory();
+		
+		FileLevels fileLevels = root.getLevelsFile(config.getFileLevelsLocation());
+		FileMessages fileMessages = root.getMessagesFile(config.getFileMessagesLocation());
+		FileScoreboard fileScoreboard = root.getScoreboardFile(config.getFileScoreboardLocation());
+		FileOptions fileOptions = root.getOptionsFile(config.getFileOptionsLocation());
+		
+		if (fileLevels.isAccessable()) {
+			this.levelManager = LevelManager.create(fileLevels);
 		}
+		
+		if (fileMessages.isAccessable()) {
+			this.messanger = Messanger.create(fileMessages);
+		}
+		
+		if (fileScoreboard.isAccessable()) {
+			this.board = GunGameScoreboard.create(fileScoreboard, this);
+		}
+		
+		if (fileOptions.isAccessable()) {
+			this.options = GameOptions.create(fileOptions);
+		}
+		
+		if (config.isAccessable()) {
+			this.locationManager = LocationManager.create(config);
+			this.config = config;
+		}
+		
+		if (this.board != null) {
+			if (this.board.isEnabled()) {
+				if (this.board.exists()) {
+					this.board.delete();
+				}
+				
+				this.board.create();
+			}
+		}
+	}
+	
+	@Override
+	public int getSize() {
+		return this.config.getMaxPlayers();
 	}
 	
 	@Override
@@ -44,7 +99,7 @@ public abstract class GunGame extends CraftGame<GunGamePlayer> {
 			msg = Main.layout().colorPrimary + "{" + Main.layout().colorSecondary + msg + Main.layout().colorPrimary + "}";
 		}
 		
-		super.playConsoleMessage(Main.layout().colorSecondary + this.settings.getUUID() + ": " + msg);
+		super.playConsoleMessage(Main.layout().colorSecondary + this.name + ": " + msg);
 	}
 	
 	@Override
@@ -64,16 +119,27 @@ public abstract class GunGame extends CraftGame<GunGamePlayer> {
 		}
 	}
 	
-	public ActionInterface getInterface() {
-		return this.actionInterface;
+	@Override
+	public void mute(boolean mute) {
+		super.mute(mute);
+		
+		if (this.config.isAccessable()) {
+			this.config.setMuted(mute);
+		} else {
+			// TODO Warning
+		}
 	}
 	
-	public Settings getSettings() {
-		return this.settings;
+	public void setEditMode(boolean value) {
+		if (this.config != null) {
+			this.config.setEditMode(value);
+		} else {
+			// TODO
+		}
 	}
 	
-	public GunGameScoreboard getScoreboard() {
-		return this.board;
+	public void initWinner(GunGamePlayer player) {
+		this.winner = player;
 	}
 	
 	public void handleDisconnect(Player player) {
@@ -95,24 +161,76 @@ public abstract class GunGame extends CraftGame<GunGamePlayer> {
 		if (this.getPlayers().size() == 0) {
 			this.playConsoleMessage(Main.layout().format("Shutdown!"));
 			
-			this.shutdown();
+			this.gameInterface.initShutdown();
 		}
 	}
 	
-	public void initWinner(GunGamePlayer player) {
-		this.winner = player;
+	public boolean isEditMode() {
+		return this.config.isEditMode();
+	}
+	
+	public boolean isReady() {
+		if (!((CraftLevelManager) this.levelManager).getFile().isReady()) {
+			return false;
+		}
+		
+		if (!((CraftMessanger) this.messanger).getFile().isReady()) {
+			return false;
+		}
+		
+		if (!((CraftGunGameScoreboard) this.board).getFile().isReady()) {
+			return false;
+		}
+		
+		if (!((CraftGameOptions) this.options).getFile().isReady()) {
+			return false;
+		}
+		
+		if (this.getLocationManager().getSpawnPointCount() < 1) {
+			return false;
+		}
+		
+		return !this.isEditMode();
+	}
+	
+	public int getMinPlayerCount() {
+		return this.config.getMinPlayers();
+	}
+	
+	public int getStartLevel() {
+		return this.config.getStartLevel();
+	}
+	
+	public String getName() {
+		return this.name;
+	}
+	
+	public GunGameScoreboard getScoreboard() {
+		return this.board;
+	}
+	
+	public Messanger getMessanger() {
+		return this.messanger;
+	}
+	
+	public LevelManager getLevelManager() {
+		return this.levelManager;
+	}
+	
+	public GameOptions getOptions() {
+		return this.options;
+	}
+	
+	public LocationManager getLocationManager() {
+		return this.locationManager;
+	}
+	
+	public GameInterface getInterface() {
+		return this.gameInterface;
 	}
 	
 	public GunGamePlayer getWinner() {
 		return this.winner;
 	}
-	
-	public abstract boolean isReady();
-	
-	public abstract void shutdown();
-	
-	public abstract void reload();
-	
-	public abstract void restart();
 	
 }
